refactor instrumentation.ts to use stricter AWS types, remove a ton of any commands (or explicitly except them) then split streaming logic out into it's own file
