+++ js/packages/openinference-instrumentation-bedrock/src/instrumentation.ts
@@ -8,12 +8,10 @@ import {
   diag,
   SpanKind,
   SpanStatusCode,
-  Span,
   context,
 } from "@opentelemetry/api";
 import {
   SemanticConventions,
-  MimeType,
   OpenInferenceSpanKind,
 } from "@arizeai/openinference-semantic-conventions";
 import { getAttributesFromContext } from "@arizeai/openinference-core";
@@ -23,13 +21,14 @@ import {
   InvokeModelWithResponseStreamCommand,
   ConverseCommand,
   BedrockRuntimeClient,
+  InvokeModelResponse,
+  ConverseResponse,
 } from "@aws-sdk/client-bedrock-runtime";
 import { extractInvokeModelRequestAttributes } from "./attributes/invoke-model-request-attributes";
 import { extractInvokeModelResponseAttributes } from "./attributes/invoke-model-response-attributes";
 import { extractConverseRequestAttributes } from "./attributes/converse-request-attributes";
 import { extractConverseResponseAttributes } from "./attributes/converse-response-attributes";
-import { setSpanAttribute } from "./attributes/attribute-helpers";
-import { isToolUseContent } from "./types/bedrock-types";
+import { consumeBedrockStreamChunks } from "./attributes/invoke-model-streaming-response-attributes";
 import { splitStream } from "@smithy/util-stream";
 
 const MODULE_NAME = "@aws-sdk/client-bedrock-runtime";
@@ -43,24 +42,6 @@ interface BedrockModuleExports {
  * Track if the Bedrock instrumentation is patched
  */
 let _isBedrockPatched = false;
-
-/**
- * Interface for streaming event data from Bedrock
- */
-interface StreamEventData {
-  type: "message_start" | "content_block_start" | "content_block_delta" | "message_delta";
-  message?: {
-    usage?: Record<string, number>;
-  };
-  content_block?: {
-    type: string;
-    text?: string;
-  };
-  delta?: {
-    text?: string;
-  };
-  usage?: Record<string, number>;
-}
 
 /**
  * Check if Bedrock instrumentation is enabled/disabled
@@ -106,11 +87,13 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
       this._wrap(
         moduleExports.BedrockRuntimeClient.prototype,
         "send",
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
         (original: any) => {
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
           return function patchedSend(this: unknown, command: any) {
             if (command?.constructor?.name === "InvokeModelCommand") {
               return instrumentation.handleInvokeModelCommand(
-                command,
+                command as InvokeModelCommand,
                 original,
                 this,
               );
@@ -121,7 +104,7 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
               "InvokeModelWithResponseStreamCommand"
             ) {
               return instrumentation.handleInvokeModelWithResponseStreamCommand(
-                command,
+                command as InvokeModelWithResponseStreamCommand,
                 original,
                 this,
               );
@@ -129,7 +112,7 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
 
             if (command?.constructor?.name === "ConverseCommand") {
               return instrumentation.handleConverseCommand(
-                command,
+                command as ConverseCommand,
                 original,
                 this,
               );
@@ -149,7 +132,9 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
 
   private handleInvokeModelCommand(
     command: InvokeModelCommand,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     original: any,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     client: any,
   ) {
     const span = this.tracer.startSpan("bedrock.invoke_model", {
@@ -170,17 +155,17 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
     extractInvokeModelRequestAttributes({ span, command });
 
     try {
-      const result = original.apply(client, [command]);
+      const result = original.apply(client, [command]) as Promise<InvokeModelResponse>;
 
       // AWS SDK v3 send() method always returns a Promise
       return result
-        .then((response: any) => {
+        .then((response: InvokeModelResponse) => {
           extractInvokeModelResponseAttributes({ span, response });
           span.setStatus({ code: SpanStatusCode.OK });
           span.end();
           return response;
         })
-        .catch((error: any) => {
+        .catch((error: Error) => {
           span.recordException(error);
           span.setStatus({
             code: SpanStatusCode.ERROR,
@@ -189,197 +174,26 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
           span.end();
           throw error;
         });
-    } catch (error: any) {
+    } catch (error) {
       // Handle errors that occur before the Promise is returned (e.g. invalid parameters)
-      span.recordException(error);
-      span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      if (error instanceof Error) {
+        span.recordException(error);
+        span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      } else {
+        span.setStatus({ code: SpanStatusCode.ERROR, message: String(error) });
+      }
       span.end();
       throw error;
     }
   }
 
-  /**
-   * Consumes AWS Bedrock streaming response chunks and extracts attributes for OpenTelemetry span.
-   *
-   * This function processes the Bedrock streaming format which consists of JSON lines
-   * containing different event types (message_start, content_block_delta, etc.).
-   * It accumulates the response content and sets appropriate semantic convention attributes
-   * on the provided span. This function is designed to run in the background without
-   * blocking the user's stream consumption.
-   *
-   * @param stream - The Bedrock response stream (AsyncIterable), typically from splitStream()
-   * @param span - The OpenTelemetry span to set attributes on
-   * @throws {Error} If critical stream processing errors occur
-   *
-   * @example
-   * ```typescript
-   * // Background processing after stream splitting
-   * const [instrumentationStream, userStream] = await splitStream(response.body);
-   * this._consumeBedrockStreamChunks(instrumentationStream, span)
-   *   .then(() => span.end())
-   *   .catch(error => { span.recordException(error); span.end(); });
-   * ```
-   */
-  /**
-   * Type guard to validate stream chunk structure
-   */
-  private isValidStreamChunk(chunk: unknown): chunk is { chunk: { bytes: Uint8Array } } {
-    return (
-      chunk !== null &&
-      typeof chunk === "object" &&
-      "chunk" in chunk &&
-      chunk.chunk !== null &&
-      typeof chunk.chunk === "object" &&
-      "bytes" in chunk.chunk &&
-      chunk.chunk.bytes instanceof Uint8Array
-    );
-  }
+  private unpatch(moduleExports: BedrockModuleExports, moduleVersion?: string) {
+    diag.debug(`Removing patch for ${MODULE_NAME}@${moduleVersion}`);
 
-  /**
-   * Type guard to validate streaming event data structure
-   */
-  private isValidStreamEventData(data: unknown): data is StreamEventData {
-    return (
-      data !== null &&
-      typeof data === "object" &&
-      "type" in data &&
-      typeof data.type === "string" &&
-      ["message_start", "content_block_start", "content_block_delta", "message_delta"].includes(data.type)
-    );
-  }
-
-  private async _consumeBedrockStreamChunks(
-    stream: AsyncIterable<unknown>,
-    span: Span,
-  ): Promise<void> {
-    let outputText = "";
-    const contentBlocks: Array<{ type: string; text?: string }> = [];
-    let usage: Record<string, number> = {};
-
-    for await (const chunk of stream) {
-      // Type guard for chunk structure
-      if (this.isValidStreamChunk(chunk)) {
-        const text = new TextDecoder().decode(chunk.chunk.bytes);
-        const lines = text.split("\n").filter((line) => line.trim());
-
-        for (const line of lines) {
-          if (line.trim()) {
-            try {
-              const rawData = JSON.parse(line);
-              if (!this.isValidStreamEventData(rawData)) {
-                continue; // Skip invalid event data
-              }
-              const data: StreamEventData = rawData;
-
-              // Handle different event types
-              if (data.type === "message_start" && data.message) {
-                usage = data.message.usage || {};
-              }
-
-              if (data.type === "content_block_start" && data.content_block) {
-                contentBlocks.push(data.content_block);
-              }
-
-              if (data.type === "content_block_delta" && data.delta?.text) {
-              // Accumulate text content
-              outputText += data.delta.text;
-
-              // Also update the content block for tool processing
-              const lastTextBlock = contentBlocks.find(
-                (block) => block.type === "text",
-              );
-              if (lastTextBlock) {
-                lastTextBlock.text =
-                  (lastTextBlock.text || "") + data.delta.text;
-              } else {
-                contentBlocks.push({
-                  type: "text",
-                  text: data.delta.text,
-                });
-              }
-            }
-
-            if (data.type === "message_delta" && data.usage) {
-              usage = { ...usage, ...data.usage };
-            }
-            } catch (parseError) {
-              // Skip malformed JSON lines silently
-              continue;
-            }
-          }
-        }
-      }
+    if (moduleExports?.BedrockRuntimeClient) {
+      this._unwrap(moduleExports.BedrockRuntimeClient.prototype, "send");
+      _isBedrockPatched = false;
     }
-
-    // Create simple output representation with only actual data we have
-    const outputValue = {
-      text: outputText || "",
-      tool_calls: contentBlocks.filter(isToolUseContent).map(content => ({
-        id: content.id,
-        name: content.name,
-        input: content.input
-      })),
-      usage: usage || {},
-      streaming: true
-    };
-
-    // Set output value as JSON
-    setSpanAttribute(
-      span,
-      SemanticConventions.OUTPUT_VALUE,
-      JSON.stringify(outputValue),
-    );
-    setSpanAttribute(span, SemanticConventions.OUTPUT_MIME_TYPE, MimeType.JSON);
-
-    // Set structured output message attributes for text content
-    if (outputText) {
-      setSpanAttribute(
-        span,
-        `${SemanticConventions.LLM_OUTPUT_MESSAGES}.0.${SemanticConventions.MESSAGE_ROLE}`,
-        "assistant",
-      );
-      setSpanAttribute(
-        span,
-        `${SemanticConventions.LLM_OUTPUT_MESSAGES}.0.${SemanticConventions.MESSAGE_CONTENT}`,
-        outputText,
-      );
-    }
-
-    // Extract tool call attributes from content blocks
-    const toolUseBlocks = contentBlocks.filter(isToolUseContent);
-    toolUseBlocks.forEach((content, toolCallIndex) => {
-      setSpanAttribute(
-        span,
-        `${SemanticConventions.LLM_OUTPUT_MESSAGES}.0.${SemanticConventions.MESSAGE_TOOL_CALLS}.${toolCallIndex}.${SemanticConventions.TOOL_CALL_FUNCTION_NAME}`,
-        content.name,
-      );
-      setSpanAttribute(
-        span,
-        `${SemanticConventions.LLM_OUTPUT_MESSAGES}.0.${SemanticConventions.MESSAGE_TOOL_CALLS}.${toolCallIndex}.${SemanticConventions.TOOL_CALL_FUNCTION_ARGUMENTS_JSON}`,
-        content.input ? JSON.stringify(content.input) : undefined,
-      );
-      setSpanAttribute(
-        span,
-        `${SemanticConventions.LLM_OUTPUT_MESSAGES}.0.${SemanticConventions.MESSAGE_TOOL_CALLS}.${toolCallIndex}.${SemanticConventions.TOOL_CALL_ID}`,
-        content.id,
-      );
-    });
-
-    // Set usage attributes directly
-    if (usage) {
-      setSpanAttribute(
-        span,
-        SemanticConventions.LLM_TOKEN_COUNT_PROMPT,
-        usage.input_tokens,
-      );
-      setSpanAttribute(
-        span,
-        SemanticConventions.LLM_TOKEN_COUNT_COMPLETION,
-        usage.output_tokens,
-      );
-    }
-
-    span.setStatus({ code: SpanStatusCode.OK });
   }
 
   /**
@@ -397,7 +211,9 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
    */
   private handleInvokeModelWithResponseStreamCommand(
     command: InvokeModelWithResponseStreamCommand,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     original: any,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     client: any,
   ) {
     const span = this.tracer.startSpan("bedrock.invoke_model", {
@@ -415,56 +231,90 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
     span.setAttributes(contextAttributes);
 
     // Extract request attributes directly onto the span
-    extractInvokeModelRequestAttributes({ span, command: command as any });
+    // Note: InvokeModelWithResponseStreamCommand has compatible input structure with InvokeModelCommand
+    extractInvokeModelRequestAttributes({ 
+      span, 
+      command: command as unknown as InvokeModelCommand 
+    });
 
     try {
-      const result = original.apply(client, [command]);
+      const result = original.apply(client, [command]) as Promise<{ body: AsyncIterable<unknown> }>;
 
       // AWS SDK v3 send() method always returns a Promise
       return result
-        .then(async (response: any) => {
+        .then(async (response: { body: AsyncIterable<unknown> }) => {
           try {
+            // Check if response.body exists before splitting
+            if (!response.body) {
+              span.setStatus({ code: SpanStatusCode.ERROR, message: "Response body is undefined" });
+              span.end();
+              return response;
+            }
+
             // Split the stream for instrumentation and user consumption
-            const [instrumentationStream, userStream] = await splitStream(
-              response.body,
-            );
+            // Note: splitStream expects a Node.js Readable or Web ReadableStream
+            const splitResult = await splitStream(
+              response.body as Parameters<typeof splitStream>[0],
+            ) as [AsyncIterable<unknown>, AsyncIterable<unknown>];
+            
+            // splitStream always returns [stream1, stream2], never null
+            const instrumentationStream = splitResult[0];
+            const userStream = splitResult[1];
 
             // Process instrumentation stream in background (non-blocking)
-            this._consumeBedrockStreamChunks(instrumentationStream, span)
-              .then(() => {
-                span.end();
+            if (instrumentationStream) {
+              // @ts-expect-error - splitStream from @smithy/util-stream always returns non-null streams
+              consumeBedrockStreamChunks({ 
+                stream: instrumentationStream, 
+                span 
               })
-              .catch((streamError: any) => {
-                span.recordException(streamError);
-                span.setStatus({
-                  code: SpanStatusCode.ERROR,
-                  message: streamError.message,
+                .then(() => {
+                  span.end();
+                })
+                .catch((streamError: Error) => {
+                  span.recordException(streamError);
+                  span.setStatus({
+                    code: SpanStatusCode.ERROR,
+                    message: streamError.message,
+                  });
+                  span.end();
                 });
-                span.end();
-              });
+            } else {
+              // Fallback: end span if no stream available
+              span.end();
+            }
 
             // Return response with user stream immediately
             return { ...response, body: userStream };
-          } catch (splitError: any) {
+          } catch (splitError) {
             // If stream splitting fails, fall back to original behavior
+            const errorMessage = splitError instanceof Error ? splitError.message : String(splitError);
             diag.warn(
               "Stream splitting failed, falling back to direct consumption:",
-              splitError,
+              errorMessage,
             );
             try {
-              await this._consumeBedrockStreamChunks(response.body, span);
-            } catch (streamError: any) {
-              span.recordException(streamError);
+              if (response.body) {
+                await consumeBedrockStreamChunks({ 
+                  stream: response.body, 
+                  span 
+                });
+              }
+            } catch (streamError) {
+              const streamErrorMessage = streamError instanceof Error ? streamError.message : String(streamError);
+              if (streamError instanceof Error) {
+                span.recordException(streamError);
+              }
               span.setStatus({
                 code: SpanStatusCode.ERROR,
-                message: streamError.message,
+                message: streamErrorMessage,
               });
             }
             span.end();
             return response;
           }
         })
-        .catch((error: any) => {
+        .catch((error: Error) => {
           span.recordException(error);
           span.setStatus({
             code: SpanStatusCode.ERROR,
@@ -473,10 +323,14 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
           span.end();
           throw error;
         });
-    } catch (error: any) {
+    } catch (error) {
       // Handle errors that occur before the Promise is returned (e.g. invalid parameters)
-      span.recordException(error);
-      span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      if (error instanceof Error) {
+        span.recordException(error);
+        span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      } else {
+        span.setStatus({ code: SpanStatusCode.ERROR, message: String(error) });
+      }
       span.end();
       throw error;
     }
@@ -484,7 +338,9 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
 
   private handleConverseCommand(
     command: ConverseCommand,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     original: any,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     client: any,
   ) {
     const span = this.tracer.startSpan("bedrock.converse", {
@@ -505,18 +361,18 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
     extractConverseRequestAttributes({ span, command });
 
     try {
-      const result = original.apply(client, [command]);
+      const result = original.apply(client, [command]) as Promise<ConverseResponse>;
 
       // AWS SDK v3 send() method always returns a Promise
       return result
-        .then((response: any) => {
+        .then((response: ConverseResponse) => {
           // Extract response attributes
-          extractConverseResponseAttributes({ span, response: response as any });
+          extractConverseResponseAttributes({ span, response });
           span.setStatus({ code: SpanStatusCode.OK });
           span.end();
           return response;
         })
-        .catch((error: any) => {
+        .catch((error: Error) => {
           span.recordException(error);
           span.setStatus({
             code: SpanStatusCode.ERROR,
@@ -525,21 +381,16 @@ export class BedrockInstrumentation extends InstrumentationBase<BedrockInstrumen
           span.end();
           throw error;
         });
-    } catch (error: any) {
+    } catch (error) {
       // Handle errors that occur before the Promise is returned (e.g. invalid parameters)
-      span.recordException(error);
-      span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      if (error instanceof Error) {
+        span.recordException(error);
+        span.setStatus({ code: SpanStatusCode.ERROR, message: error.message });
+      } else {
+        span.setStatus({ code: SpanStatusCode.ERROR, message: String(error) });
+      }
       span.end();
       throw error;
-    }
-  }
-
-  private unpatch(moduleExports: BedrockModuleExports, moduleVersion?: string) {
-    diag.debug(`Removing patch for ${MODULE_NAME}@${moduleVersion}`);
-
-    if (moduleExports?.BedrockRuntimeClient) {
-      this._unwrap(moduleExports.BedrockRuntimeClient.prototype, "send");
-      _isBedrockPatched = false;
     }
   }
 }
